+++ CHANGELOG.md
@@ -3,8 +3,12 @@
 Änderungen
 ----------
 
-v1.1.1 (??.??.2016)
+v1.1.2 (??.??.2016)
+
+v1.1.1 (17.08.2016)
  * new Tx_Rnbase_Utility_Logger added
+ * hide show links in decorator fixed
+ * new table utility for be modules
 
 v1.1.0 (11.08.2016)
  * dirty state after setProperty([]) should be dirty, not clean